Add timeout to subprocess.communicate() calls
There's cases where amulet would hang for long periods of time (hours
and hours). It seems this is due to subprocess.communicate() calls
blocking with hung juju processes (see GitHub #184). This updates
calls to use a hard timeout.

Unfortunately, timeout was introduced in Python 3.3. so this breaks
for older versions of Python.